Comments Route resource

+++ app/Http/Controllers/CommentController.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use App\Models\Comment;
 use Illuminate\Http\Request;
 
 class CommentController extends Controller
 {
     public function destroy(Comment $comment){
+        $this->authorize('delete', $comment);
         $comment->delete();
         return back();
     }
